Split MessageMaker into smaller files
